+++ job.h
@@ -5,12 +5,12 @@ typedef struct cs537_Job /*Tag for the struct*/
 { 
     char * cmd_name;
     int arrival_time;
-    int CPU_time;
+    double CPU_time;
     int time_remaining;
     double time_running;
-    int IO_count;
-    int IOOperations;
-    int IO_interval;
+    double IO_count;
+    double IOOperations;
+    double IO_interval;
     int timeTillNextIO;
 }Job,*JobP;/*Actual struct name*/
 #endif
